requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
aspose-words==25.2.0
async-timeout==5.0.1
attrs==25.1.0
Camelot==12.6.29
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
decorator==5.1.1
distro==1.9.0
docx2pdf==0.1.8
Elixir==0.7.1
exceptiongroup==1.2.2
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.0.2
openai==1.61.1
packaging==24.2
pbr==6.1.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
plum-dispatch==1.7.4
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pymupdf==1.25.3
pypdf==5.3.0
pypdfium2==4.30.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-multipart==0.0.20
pywin32==308
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
Spire.Doc==13.1.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.13.0
sqlparse==0.5.3
starlette==0.45.3
Tempita==0.6.0
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
xlrd==0.7.1
xlwt==0.7.2
yarl==1.18.3
